requirements.txt
flask
flask-socketio
python-dotenv
requests
Pillow
opencv-python
ximilar-client